ts/examples/docuProc/requirements.txt
# Copyright (c) Microsoft Corporation.
# Licensed under the MIT License.

pdfplumber
pymupdf
pytesseract
Pillow